Fix ImmData grain index uint16 truncation in egress RMA protocol
The ImmDataGrain constructor previously accepted a uint64_t grain index and
silently truncated it to uint16_t for the ring buffer slot field. A monotonic
grain index wraps after 65 536 grains (~18 minutes at 60 fps), after which the
receiver decodes the wrong ring buffer slot and reads stale grain data,
causing latency spikes.

Change the constructor parameter to uint16_t so the ring buffer slot must be
computed explicitly by the caller (as grainIndex % grainCount). The call site
in RMAGrainEgressProtocol::transferGrain already computes this modulo; it now
casts the result explicitly. This makes the contract clear and prevents
accidental silent truncation of raw 64-bit indices.

Add unit tests covering pack/unpack round-trip, max slot value, and zero.

+++ lib/fabrics/ofi/src/internal/ProtocolEgressRMA.cpp
@@ -35,7 +35,7 @@ namespace mxl::lib::fabrics::ofi
 
         auto const localRegion = _localRegions[localIndex % _localRegions.size()].sub(localOffset, localSize);
         auto const remoteRegion = _remoteInfo.remoteRegions[remoteIndex % _remoteInfo.remoteRegions.size()].sub(remoteOffset, remoteSize);
-        auto const remoteSlot = remoteIndex % _remoteInfo.remoteRegions.size();
+        auto const remoteSlot = static_cast<std::uint16_t>(remoteIndex % _remoteInfo.remoteRegions.size());
 
         _pending += ep.write(_token, localRegion, remoteRegion, destAddr, ImmDataGrain{remoteSlot, sliceRange.end()}.data());
     }
